Move MinMaxTimestampTransformer to core-metrics
