+++ core/metrics-core-service/src/main/java/org/hawkular/metrics/core/service/transformers/MinMaxTimestampTransformer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.api.jaxrs.handler.transformer;
+package org.hawkular.metrics.core.service.transformers;
 
 import org.hawkular.metrics.core.service.MetricsService;
 import org.hawkular.metrics.core.service.Order;
